Task12 fixed tests
